backend/requirements.txt
# Core dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0

# HTTP client for LLM API
httpx>=0.27.0

# WebSocket client for ElevenLabs TTS
websockets>=13.0

# No LiveKit dependencies - removed as requested

